requirements.txt
fastapi
uvicorn
wheel
loguru
python-dotenv
unidecode
redis
pymongo
numpy
pydantic
openai
requests
email-validator
